requirements_cli.txt
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.24.3
python-whois==0.8.0
tldextract==5.0.1

